♻️ Refactor observers into core, add serializer
