+++ core/harambe_core/observer/json_line_observer.py
@@ -1,0 +1,8 @@
+import json
+
+from harambe_core.observer.serialization_observer import SerializationObserver
+
+
+class JsonLineObserver(SerializationObserver):
+    def __int__(self):
+        super().__init__(sink=print, serializer=json.dumps)
